requirements.txt
python-telegram-bot
geopy
requests